requirements.txt
aiohttp~=3.7.4.post0
aiosqlite~=0.17.0
python-dotenv~=0.18.0